refactor(task-queue) renamed initalizer.js to initialize.js
